requirements.txt
torch>=1.13.1
torchaudio>=0.13.1
librosa>=0.9.0
matplotlib>=3.7.0
scikit-learn>=1.2.1
tqdm>=4.64.1

jupyter
soundfile
streamlit
audio-recorder-streamlit
numpy
# audiomentations  # Opsiyonel: Yüklemek için Visual C++ Build Tools gerekli
moviepy

# Database
psycopg2-binary
pgvector
python-dotenv

# VAD
webrtcvad-wheels
